SpecRunner now ignores contexts that have been annotated with org.junit.Ignore.
I added this because the common convention right now for temporarily ignoring a context seems to be to make it package-protected, which is rather implicit. Explicitly marking a context @Ignored is much better, in my opinion.

Other possible mechanisms for ignoring a context could be a marker interface. We could also provide a JDave-specific annotation, "jdave.Ignored" (although that might lead to name-clashes and ugly fully-qualified references).